+++ packages/solidart/lib/src/core/resource.dart
@@ -148,15 +148,15 @@ class Resource<ResultType> extends Signal<ResourceValue<ResultType>> {
   void _initialize() {
     // React to the [source], if provided.
     if (fetcher != null && source != null) {
-      source!.addListener(fetch);
+      source!.addListener(refetch);
       source!.onDispose(() => source!.removeListener(fetch));
     }
     // React the the [stream], if provided
     if (stream != null) {
       _streamSubscription = stream!.listen((data) {
-        value = ResourceValue.ready(data);
+        value = ResourceValue<ResultType>.ready(data);
       }, onError: (error, stackTrace) {
-        value = ResourceValue.error(error, stackTrace: stackTrace);
+        value = ResourceValue<ResultType>.error(error, stackTrace: stackTrace);
       });
     }
   }
@@ -167,12 +167,14 @@ class Resource<ResultType> extends Signal<ResourceValue<ResultType>> {
   /// performed by [ResourceBuilder].
   Future<void> fetch() async {
     assert(fetcher != null, "You are trying to fetch, but fetcher is null");
+    assert(value is ResourceUnresolved<ResultType>,
+        "Cannot fetch a resource that is already resolved, use 'refetch' instead");
     try {
-      value = const ResourceValue.loading();
+      value = ResourceValue<ResultType>.loading();
       final result = await fetcher!();
-      value = ResourceValue.ready(result);
+      value = ResourceValue<ResultType>.ready(result);
     } catch (e, s) {
-      value = ResourceValue.error(e, stackTrace: s);
+      value = ResourceValue<ResultType>.error(e, stackTrace: s);
     }
   }
 
@@ -180,18 +182,18 @@ class Resource<ResultType> extends Signal<ResourceValue<ResultType>> {
   Future<void> refetch() async {
     assert(fetcher != null, "You are trying to refetch, but fetcher is null");
     try {
-      if (value is ResourceReady) {
+      if (value is ResourceReady<ResultType>) {
         update(
           (value) =>
               (value as ResourceReady<ResultType>).copyWith(refreshing: true),
         );
       } else {
-        value = const ResourceValue.loading();
+        value = ResourceValue<ResultType>.loading();
       }
       final result = await fetcher!();
-      value = ResourceValue.ready(result);
+      value = ResourceValue<ResultType>.ready(result);
     } catch (e, s) {
-      value = ResourceValue.error(e, stackTrace: s);
+      value = ResourceValue<ResultType>.error(e, stackTrace: s);
     }
   }
 
@@ -200,6 +202,10 @@ class Resource<ResultType> extends Signal<ResourceValue<ResultType>> {
     _streamSubscription?.cancel();
     super.dispose();
   }
+
+  @override
+  String toString() =>
+      '''Resource<$ResultType>(value: $value, previousValue: $previousValue, options; $options)''';
 }
 
 @sealed
@@ -393,13 +399,13 @@ class ResourceUnresolved<T> implements ResourceValue<T> {
 
 // coverage:ignore-start
 extension ResourceExtensions<T> on ResourceValue<T> {
-  /// Indicates if the resoruce is loading.
+  /// Indicates if the resource is loading.
   bool get isLoading => this is ResourceLoading<T>;
 
-  /// Indicates if the resoruce has an error.
+  /// Indicates if the resource has an error.
   bool get hasError => this is ResourceError<T>;
 
-  /// Indicates if the resoruce is ready.
+  /// Indicates if the resource is ready.
   bool get isReady => this is ResourceReady<T>;
 
   /// Upcast [ResourceValue] into a [ResourceReady], or return null if the